fix: keep track of the last distribution
`stakingRewards.lastUpdatedTime()` is not reliable because it is updated not only by `notifyRewardAmount`, but also by `stake` and `withdraw`.

The catch though is that the actual calculation depends on the underlying vesting stream, not the actual distribution. If the `vestId` changes, we need to reset the value of the last distribution time.